requirements.txt
# AI Workspace Requirements
# Multi-Domain Knowledge Base System

# Core AI/ML frameworks
crewai>=0.28.0
crewai-tools>=0.12.0
mem0ai>=0.1.0
langchain>=0.1.0
anthropic>=0.18.0

# Vector Databases
qdrant-client>=1.7.0
chromadb>=0.4.0

# Embeddings & LLMs
openai>=1.12.0
cohere>=4.47

# Search & Research APIs
tavily-python>=0.3.0
exa-py>=1.0.0
firecrawl-py>=0.0.4

# Academic Research
arxiv>=2.1.0
semanticscholar>=0.8.0
biopython>=1.83  # For PubMed

# Social Media
praw>=7.7.0  # Reddit
tweepy>=4.14.0  # Twitter/X

# Web Scraping & Data
beautifulsoup4>=4.12.0
requests>=2.31.0
feedparser>=6.0.10  # RSS feeds
newspaper3k>=0.2.8  # Article extraction
youtube-transcript-api>=0.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
supabase>=2.0.0

# API Server
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
colorama>=0.4.6
tqdm>=4.65.0
rich>=13.7.0

# Optional: Local LLMs (if needed)
# transformers>=4.36.0
# torch>=2.1.0
